requirements.txt
alembic==1.4.0
amqp==2.5.2
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
billiard==3.6.3.0
blinker==1.4
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cssselect==1.1.0
distlib==0.3.0
entrypoints==0.3
feedfinder2==0.0.4
feedparser==5.2.1
filelock==3.0.12
flake8==3.7.9
Flask==1.1.1
Flask-Dance==3.0.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
httplib2==0.17.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.1
kombu==4.6.7
lazy-object-proxy==1.4.3
lxml==4.5.0
Mako==1.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
newspaper3k==0.2.8
nltk==3.4.5
oauth2==1.9.0.post1
oauthlib==3.1.0
packaging==20.1
Pillow==7.0.0
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyJWT==1.4.2
pylint==2.4.4
pylint-flask-sqlalchemy==0.2.0
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3
requests==2.23.0
requests-file==1.4.3
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.13
SQLAlchemy-Utils==0.36.1
tinysegmenter==0.3
tldextract==2.2.2
toml==0.10.0
tox==3.14.5
urllib3==1.25.8
URLObject==2.4.3
vine==1.3.0
virtualenv==20.0.7
wcwidth==0.1.8
Werkzeug==1.0.0
wrapt==1.11.2
WTForms==2.2.1
